TravelLocationsMapVC nearly done, some code clean up and slight refactor happened. TODO comment was created to deselect pins after navigating back to the TravelLocationMap Controller.
